requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# LLM Providers
openai==1.6.1
anthropic==0.8.1
dashscope==1.14.1  # 阿里云通义千问
zhipuai==2.0.1  # 智谱AI

# Vector Databases
chromadb==0.4.22
faiss-cpu==1.7.4
qdrant-client==1.7.0

# Embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2

# Database Connectors
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
httpx==0.26.0
tenacity==8.2.3

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0
